-vvh rsync for logs
After some testing it looks like -v does not contain enough details for
my taste, and -vvv has too much (9x more lines that -vv) and that starts
to bog down file size. Also removed --protect-args in line 30, because
with the commit that fixes $dtv, it showed to be unnecessary.