Adds note on vimconfig.sh error message
I tried really hard to find a way to get vim to start without loading
a colorscheme, but couldn't find anything at all. Because molokai
is called for when vim is started the first time (before the colorscheme
is actually installed), vim throws an error message, and you need to
manually hit start.