.vimrc better for scripting; relative airline path
Adding 'silent!' to the colorscheme line. Previously, running
vimconfig.sh on a fresh vim throws an error message opening vim for the
first time to install plugins since the colorscheme hasn't been
installed. This error requires user-input to proceed and is thus
undesirable. The prepending of 'silent!' to the colorscheme line should
eliminate this error message.

Using full filepaths in airline turned out to be cumbersome and just
truncated other parts of the line.